+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * Better error message when invalid values are passed to generated
     Python entry points.
 
+  * `futhark-doc` now ignores files whose doc comment is the word
+    "ignore".
+
 ### Removed
 
 ### Changed